+++ cinelerra/filelist.h
@@ -47,7 +47,7 @@ public:
 	int write_list_header();
 	int write_frames(VFrame ***frames, int len);
 	VFrame* read_frame(int use_alpha, int use_float);
-	virtual int get_memory_usage();
+	virtual int64_t get_memory_usage();
 // Get the total writer units for calculating memory usage
 	int get_units();
 // Get a writer unit for retrieving temporary usage.
